split alterBuiltManifest.ts into 2 files
