+++ app/scripts/lib/manifestFlags.ts
@@ -1,3 +1,5 @@
+import browser from 'webextension-polyfill';
+
 export type ManifestFlags = {
   circleci?: {
     enabled: boolean;
@@ -10,11 +12,20 @@ export type ManifestFlags = {
   doNotForceSentryForThisTest?: boolean;
 };
 
+// eslint-disable-next-line @typescript-eslint/consistent-type-definitions -- you can't extend a type, we want this to be an interface
+interface WebExtensionManifestWithFlags
+  extends browser.Manifest.WebExtensionManifest {
+  _flags?: ManifestFlags;
+}
+
 /**
- * Get the runtime flags that were placed in manifest.json by alterBuiltManifest.ts
+ * Get the runtime flags that were placed in manifest.json by manifest-flag-mocha-hooks.ts
  *
  * @returns flags if they exist, otherwise an empty object
  */
 export function getManifestFlags(): ManifestFlags {
-  return chrome.runtime.getManifest()._flags || {};
+  return (
+    (browser.runtime.getManifest() as WebExtensionManifestWithFlags)._flags ||
+    {}
+  );
 }
