+++ test/e2e/manifest-flag-mocha-hooks.ts
@@ -14,14 +14,12 @@
  */
 import fs from 'fs';
 import { hasProperty } from '@metamask/utils';
-import { ManifestFlags } from '../../app/scripts/lib/manifestFlags';
-
-const folder = `dist/${process.env.SELENIUM_BROWSER}`;
+import { folder } from './set-manifest-flags';
 
 // Global beforeEach hook to backup the manifest.json file
 if (typeof beforeEach === 'function' && process.env.SELENIUM_BROWSER) {
   beforeEach(() => {
-    console.debug('alterBuiltManifest.ts -- beforeEach hook');
+    console.debug('manifest-flag-mocha-hooks.ts -- beforeEach hook');
 
     restoreBackupManifest();
 
@@ -32,7 +30,7 @@ if (typeof beforeEach === 'function' && process.env.SELENIUM_BROWSER) {
 // Global afterEach hook to restore the backup manifest
 if (typeof afterEach === 'function' && process.env.SELENIUM_BROWSER) {
   afterEach(() => {
-    console.debug('alterBuiltManifest.ts -- afterEach hook');
+    console.debug('manifest-flag-mocha-hooks.ts -- afterEach hook');
 
     // create manifest.altered.json
     backupManifest('altered');
@@ -73,37 +71,4 @@ function restoreBackupManifest() {
       }
     }
   }
-}
-
-function parseIntOrUndefined(value: string | undefined): number | undefined {
-  return value ? parseInt(value, 10) : undefined;
-}
-
-/**
- * Alter the manifest with CircleCI environment variables and custom flags
- *
- * @param flags - Custom flags to set
- * @param flags.circleci - This will usually come in as undefined, and be set in this function
- */
-export function setManifestFlags(flags: ManifestFlags = {}) {
-  if (process.env.CIRCLECI) {
-    flags.circleci = {
-      enabled: true,
-      branch: process.env.CIRCLE_BRANCH,
-      buildNum: parseIntOrUndefined(process.env.CIRCLE_BUILD_NUM),
-      job: process.env.CIRCLE_JOB,
-      nodeIndex: parseIntOrUndefined(process.env.CIRCLE_NODE_INDEX),
-      prNumber: parseIntOrUndefined(
-        process.env.CIRCLE_PULL_REQUEST?.split('/').pop(), // The CIRCLE_PR_NUMBER variable is only available on forked Pull Requests
-      ),
-    };
-  }
-
-  const manifest = JSON.parse(
-    fs.readFileSync(`${folder}/manifest.json`).toString(),
-  );
-
-  manifest._flags = flags;
-
-  fs.writeFileSync(`${folder}/manifest.json`, JSON.stringify(manifest));
 }
