+++ test/e2e/set-manifest-flags.ts
@@ -1,0 +1,32 @@
+import fs from 'fs';
+import { ManifestFlags } from '../../app/scripts/lib/manifestFlags';
+
+export const folder = `dist/${process.env.SELENIUM_BROWSER}`;
+
+function parseIntOrUndefined(value: string | undefined): number | undefined {
+  return value ? parseInt(value, 10) : undefined;
+}
+
+// Alter the manifest with CircleCI environment variables and custom flags
+export function setManifestFlags(flags: ManifestFlags = {}) {
+  if (process.env.CIRCLECI) {
+    flags.circleci = {
+      enabled: true,
+      branch: process.env.CIRCLE_BRANCH,
+      buildNum: parseIntOrUndefined(process.env.CIRCLE_BUILD_NUM),
+      job: process.env.CIRCLE_JOB,
+      nodeIndex: parseIntOrUndefined(process.env.CIRCLE_NODE_INDEX),
+      prNumber: parseIntOrUndefined(
+        process.env.CIRCLE_PULL_REQUEST?.split('/').pop(), // The CIRCLE_PR_NUMBER variable is only available on forked Pull Requests
+      ),
+    };
+  }
+
+  const manifest = JSON.parse(
+    fs.readFileSync(`${folder}/manifest.json`).toString(),
+  );
+
+  manifest._flags = flags;
+
+  fs.writeFileSync(`${folder}/manifest.json`, JSON.stringify(manifest));
+}
